Extract product redirect policy
